Added flexibility for choosing datatype of resultset
